mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/HandlerMappingRegister.java
package com.interface21.webmvc.servlet.mvc.tobe;

import jakarta.servlet.http.HttpServletRequest;

import java.util.ArrayList;
import java.util.List;
import java.util.Objects;

public class HandlerMappingRegister {

    private final List<HandlerMapping> handlerMappings = new ArrayList<>();

    public void addHandlerMapping(HandlerMapping handlerMapping) {
        handlerMappings.add(handlerMapping);
    }

    public Object getHandler(HttpServletRequest request) {
        return handlerMappings.stream()
                .map(handlerMapping -> handlerMapping.getHandler(request))
                .filter(Objects::nonNull)
                .findFirst()
                .orElseThrow(() -> new IllegalArgumentException(
                        String.format("요청 URI 또는 method와 일치하는 핸들러를 찾을 수 없습니다. 요청된 URI: \"%s\", method: \"%s\"", request.getRequestURI(), request.getMethod()))
                );
    }
}
